+++ SpaceBattle.Lib/ActionCommand.cs
@@ -1,0 +1,15 @@
+namespace SpaceBattle.Lib;
+
+public class ActionCommand : ICommand
+{
+    private readonly Action _action;
+    public ActionCommand(Action action)
+    {
+        _action = action;
+    }
+
+    public void Execute()
+    {
+        _action();
+    }
+}
